add extracted data component
