+++ src/components/list/editList.vue
@@ -913,101 +913,17 @@
               :methAssessments="meth_assessments"
               :refsWithTitle="refsWithTitle"></table-meth-assessments>
 
-            <div
-              class="mt-3"
-              v-if="show.selected.includes('ed')">
-              <a name="extracted-data"></a>
-              <videoHelp txt="Extracted data" tag="h3-extracted-data" urlId="450836795" :warning="ui.adequacy.extracted_data.display_warning"></videoHelp>
-              <p class="d-print-none font-weight-light">
-                It is here that you enter the data extracted from included studies that support this review finding. This data is needed to make a GRADE-CERQual assessment.
-              </p>
-              <template v-if="extracted_data.fields.length">
-                <bc-filters
-                  v-if="mode==='edit'"
-                  class="d-print-none"
-                  idname="extracted-data-filter"
-                  :tableSettings="extracted_data_table_settings"
-                  type="extracted_data"
-                  :fields="mode_print_fieldsObj"
-                  :items="extracted_data.items">
-                </bc-filters>
-                <b-table
-                  class="toDoc extracted-data-table"
-                  :id="(mode==='view') ? 'extracted-view' : 'extracted'"
-                  responsive
-                  head-variant="light"
-                  outlined
-                  :filter="extracted_data_table_settings.filter"
-                  :fields="(mode==='view') ? mode_print_fieldsObj : extracted_data.fieldsObj"
-                  :items="extracted_data.items"
-                  :current-page="extracted_data_table_settings.currentPage">
-                  <template v-slot:cell(authors)="data">
-                    <span v-b-tooltip.hover :title="getReferenceInfo(data.item.ref_id)">{{data.item.authors}}</span>
-                  </template>
-                  <template
-                    v-if="mode==='edit' && checkPermissions(list.organization)"
-                    v-slot:cell(actions)="data">
-                    <b-button
-                      class="d-print-none"
-                      @click="openModalExtractedDataEditDataItem(data)"
-                      variant="outline-success">
-                      <font-awesome-icon
-                        icon="edit"
-                        :title="$t('Edit')" />
-                    </b-button>
-                    <b-button
-                      class="d-print-none"
-                      @click="openModalExtractedDataRemoveDataItem(data)"
-                      variant="outline-danger">
-                      <font-awesome-icon
-                        icon="trash"
-                        :title="$t('Remove')" />
-                    </b-button>
-                  </template>
-                </b-table>
-                <b-modal
-                  id="modal-extracted-data-remove-data-item"
-                  ref="modal-extracted-data-remove-data-item"
-                  title="Remove data content"
-                  @ok="extractedDataRemoveDataItem"
-                  ok-variant="outline-success"
-                  cancel-variant="outline-secondary">
-                  <p>Are you sure you want to delete all the content for this row?</p>
-                </b-modal>
-                <b-modal
-                  size="xl"
-                  title="Edit data"
-                  id="modal-extracted-data-data"
-                  ref="modal-extracted-data-data"
-                  @ok="saveDataExtractedData"
-                  cancel-variant="outline-secondary"
-                  ok-variant="outline-success"
-                  ok-title="Save">
-                  <b-form-group
-                    v-for="(field, index) in buffer_extracted_data.fields"
-                    :key="index"
-                    :id="`label-field-${index}`"
-                    :label="(field.key === 'column_0') ? 'Add the extracted data from this study that supports the review finding' : ''"
-                    :label-for="`input-field-${index}`">
-                    <b-form-textarea
-                      :id="`input-field-${index}`"
-                      v-if="field.key !== 'ref_id' && field.key !== 'authors'"
-                      v-model="buffer_extracted_data_items[field.key]"
-                      rows="6"
-                      max-rows="100"></b-form-textarea>
-                  </b-form-group>
-                </b-modal>
-
-                <back-to-top></back-to-top>
-              </template>
-              <!--
-              <template v-else>
-                <p class="d-print-none font-weight-light">
-                  To create or make changes to the column headings for this table, do so in the <b-link :to="`/organization/${list.organization}/project/${list.project_id}#My-Data`">My Data</b-link> section of iSoQ, once your headings are created you will be able to add the Extracted Data here.
-                </p>
-              </template>
-              -->
-            </div>
+            <table-extracted-data
+              :ui="ui"
+              :show="show"
+              :mode="mode"
+              :list="list"
+              :permission="checkPermissions(list.organization)"
+              :extractedData="extracted_data"
+              :modePrintFieldObject="mode_print_fieldsObj"
+              :refsWithTitle="refsWithTitle"
+              @printErrors="printErrors"
+              @getExtractedData="getExtractedData"></table-extracted-data>
 
             <template v-if="Object.prototype.hasOwnProperty.call(this.project, 'license_type')">
               <div class="mt-5">
@@ -1035,6 +951,7 @@ const editListActionButtons = () => import('./editListActionButtons.vue')
 const editListEvidenceProfile = () => import('./editListEvidenceProfile.vue')
 const editListCharsOfStudies = () => import('./editListCharsOfStudies.vue')
 const editListMethAssessments = () => import('./editListMethAssessments.vue')
+const editListExtractedData = () => import('./editListExtractedData.vue')
 
 export default {
   components: {
@@ -1048,7 +965,8 @@ export default {
     'edit-list-actions-buttons': editListActionButtons,
     'evidence-profile-table': editListEvidenceProfile,
     'table-chars-of-studies': editListCharsOfStudies,
-    'table-meth-assessments': editListMethAssessments
+    'table-meth-assessments': editListMethAssessments,
+    'table-extracted-data': editListExtractedData
   },
   data () {
     return {
@@ -1092,13 +1010,6 @@ export default {
         perPage: 10,
         pageOptions: [10, 50, 100],
         isBusy: false
-      },
-      extracted_data_table_settings: {
-        filter: '',
-        totalRows: 1,
-        currentPage: 1,
-        perPage: 10,
-        pageOptions: [10, 50, 100]
       },
       show: {
         selected: ['cs', 'ma', 'ed'],
@@ -1146,12 +1057,6 @@ export default {
         relevance: {option: null, explanation: '', notes: '', title: ''},
         cerqual: {option: null, explanation: '', notes: '', title: ''}
       },
-      buffer_extracted_data: {
-        fields: [],
-        items: [],
-        id: null
-      },
-      buffer_extracted_data_items: {},
       list: {
         id: '',
         title: '',
@@ -1367,7 +1272,7 @@ export default {
               items: _items
             }
             axios.patch(`/api/isoqf_extracted_data/${response.data[0].id}`, params)
-              .then((response) => {
+              .then(() => {
                 this.getExtractedData()
               })
           }
@@ -1859,62 +1764,6 @@ export default {
           this.printErrors(error)
         })
     },
-    saveDataExtractedData: function () {
-      let _item = JSON.parse(JSON.stringify(this.buffer_extracted_data_items))
-      let _originalItems = JSON.parse(JSON.stringify(this.extracted_data.original_items))
-
-      for (let index in _originalItems) {
-        if (_item.ref_id === _originalItems[index].ref_id) {
-          _originalItems[index] = {
-            'authors': _item.authors,
-            'column_0': _item.column_0,
-            'ref_id': _item.ref_id
-          }
-        }
-      }
-
-      let params = {
-        organization: this.list.organization,
-        list_id: this.$route.params.id,
-        items: _originalItems
-      }
-      axios.patch(`/api/isoqf_extracted_data/${this.extracted_data.id}`, params)
-        .then((response) => {
-          this.getExtractedData()
-          this.buffer_extracted_data = {fields: [], items: [], id: null}
-          this.buffer_extracted_data_items = {}
-        })
-        .catch((error) => {
-          this.printErrors(error)
-        })
-    },
-    openModalExtractedDataRemoveDataItem: function (data) {
-      this.buffer_extracted_data.remove_index_item = data.index
-      this.$refs['modal-extracted-data-remove-data-item'].show()
-    },
-    extractedDataRemoveDataItem: function () {
-      let items = JSON.parse(JSON.stringify(this.extracted_data.items))
-      const item = items[this.buffer_extracted_data.remove_index_item]
-      let newItem = { 'ref_id': item.ref_id, 'authors': item.authors, 'column_0': '' }
-      items[this.buffer_extracted_data.remove_index_item] = newItem
-
-      axios.patch(`/api/isoqf_extracted_data/${this.extracted_data.id}`, {items: items})
-        .then((response) => {
-          this.getExtractedData()
-          delete this.buffer_extracted_data.remove_index_item
-        })
-        .catch((error) => {
-          this.printErrors(error)
-        })
-    },
-    openModalExtractedDataEditDataItem: function (data) {
-      this.extracted_data.edit_index_item = data.index
-      this.buffer_extracted_data.fields = JSON.parse(JSON.stringify(this.extracted_data.fields))
-      this.buffer_extracted_data.fields.splice(this.buffer_extracted_data.fields.length - 1, 1)
-      this.buffer_extracted_data_items = JSON.parse(JSON.stringify(this.extracted_data.items[data.index]))
-      this.$refs['modal-extracted-data-data'].show()
-    },
-
     printErrors: function (error) {
       if (error.response) {
         // The request was made and the server responded with a status code
